Implement clean architecture structure layout, delete unused Shop screen and buttons in MainMenuScreen.
