+++ app/src/main/java/pl/kazoroo/dices/DiceApplication.kt
@@ -1,17 +1,17 @@
-package pl.kazoroo.dices.data
+package pl.kazoroo.dices
 
 import android.app.Application
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
-
-private const val LAYOUT_PREFERENCE_NAME = "layout_preferences"
-private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(
-        name = LAYOUT_PREFERENCE_NAME
-)
+import pl.kazoroo.dices.data.UserPreferencesRepository
 
 class DiceApplication : Application() {
+    private val LAYOUT_PREFERENCE_NAME = "layout_preferences"
+    private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(
+        name = LAYOUT_PREFERENCE_NAME
+    )
     lateinit var userPreferencesRepository: UserPreferencesRepository
 
     override fun onCreate() {
