--- app/src/main/java/pl/kazoroo/dices/navigation/BottomNavigationItems.kt
@@ -1,17 +1,0 @@
-package pl.kazoroo.dices.navigation
-
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.ui.graphics.vector.ImageVector
-
-data class BottomNavigationItem(
-    val title: String,
-    val icon: ImageVector,
-)
-
-val items = listOf(
-        BottomNavigationItem(
-                title = "Back",
-                icon = Icons.Default.ArrowBack,
-        ),
-)
