+++ app/src/main/java/pl/kazoroo/dices/presentation/navigation/Screen.kt
@@ -1,10 +1,9 @@
-package pl.kazoroo.dices.navigation
+package pl.kazoroo.dices.presentation.navigation
 
 sealed class Screen(val route: String) {
     object MainScreen : Screen("main_screen")
     object GameScreen : Screen("game_screen")
     object SettingsScreen : Screen("settings_screen")
-    object ShopScreen : Screen("shop_screen")
 
     /**
      * Creates a string route including arguments
